requirement.txt
asgiref==3.8.1
certifi==2024.8.30
Django==4.2.16
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
typing-extensions==4.12.2
wrapt==1.16.0
